+++ src/js/controllers/purchaseProductController.js
@@ -1,0 +1,3 @@
+export class purchaseProductController {
+  constructor() {}
+}
